% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probesets.R
\name{filter_probeset}
\alias{filter_probeset}
\title{Filter out the best probeset which is supposed to be specific and/or exhibiting the highest expression variance}
\usage{
filter_probeset(data, selected_probesets)
}
\arguments{
\item{data}{A tibble containing gene_symbol, probeset_id and expression columns}

\item{selected_probesets}{A tibble generated by `select_probeset()`}
}
\description{
Filter out the best probeset which is supposed to be specific and/or exhibiting the highest expression variance
}
