% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{format_publication_date}
\alias{format_publication_date}
\title{Format the publication date found in GEO datasets}
\usage{
format_publication_date(x)
}
\arguments{
\item{x}{A character vector containing the date to be formatted}
}
\description{
Format the publication date found in GEO datasets
}
