% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{get_eset_goi}
\alias{get_eset_goi}
\title{Extract genes of interest out of an ExpressionSet object}
\usage{
get_eset_goi(eset, goi, metadata)
}
\arguments{
\item{eset}{An ExpressionSet objects.}

\item{goi}{A character vector containing the gene symbols of the genes of interest}

\item{metadata}{metadata columns to join (tidyselect). If not set will return only the expression values.}
}
\description{
A helper function to get the expression data of genes of interest out of the downloaded GEO data.
}
