% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_eset_metadata}
\alias{get_eset_metadata}
\title{Extract Metadata from ExpressionSet with tidy characteristics columns}
\usage{
get_eset_metadata(x)
}
\arguments{
\item{x}{A GEO ExpressionSet}
}
\description{
Extract Metadata from ExpressionSet with tidy characteristics columns
}
