% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{tidy_key_values}
\alias{tidy_key_values}
\title{Tidy the GEO characteristics columns}
\usage{
tidy_key_values(data, ...)
}
\arguments{
\item{data}{A tibble.}

\item{...}{columns containing the key value pairs.}
}
\description{
Tidy the GEO characteristics columns
}
\details{
Characterstics columns might contain gathered and united key value pairs in form of "key: value".
This function will separate the key value pairs and spread the variables into separate columns.
}
