% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probesets.R
\name{select_probeset}
\alias{select_probeset}
\title{Identify probeset exhibiting the highest expression variance}
\usage{
select_probeset(data)
}
\arguments{
\item{data}{A tibble containing gene_symbol, probeset_id and expression columns}
}
\description{
Identify probeset exhibiting the highest expression variance
}
